Added remaining binary formats

+++ recorder/record_getdata.go
@@ -61,17 +61,18 @@ func (gr *GetDataRecord) String() string {
 
 func (gr *GetDataRecord) Bytes() []byte {
 	buf := new(bytes.Buffer)
-	binary.Write(buf, binary.LittleEndian, ParseCommand(gr.cmd))
-	binary.Write(buf, binary.LittleEndian, gr.stamp.UnixNano())
-	binary.Write(buf, binary.LittleEndian, gr.ra.IP.To16())
-	binary.Write(buf, binary.LittleEndian, uint16(gr.ra.Port))
-	binary.Write(buf, binary.LittleEndian, gr.la.IP.To16())
-	binary.Write(buf, binary.LittleEndian, uint16(gr.la.Port))
-	binary.Write(buf, binary.LittleEndian, len(gr.items))
+	binary.Write(buf, binary.LittleEndian, ParseCommand(gr.cmd))  //  1
+	binary.Write(buf, binary.LittleEndian, gr.stamp.UnixNano())   //  8
+	binary.Write(buf, binary.LittleEndian, gr.ra.IP.To16())       // 16
+	binary.Write(buf, binary.LittleEndian, uint16(gr.ra.Port))    //  2
+	binary.Write(buf, binary.LittleEndian, gr.la.IP.To16())       // 16
+	binary.Write(buf, binary.LittleEndian, uint16(gr.la.Port))    //  2
+	binary.Write(buf, binary.LittleEndian, uint16(len(gr.items))) //  2
 
-	for _, item := range gr.items {
-		binary.Write(buf, binary.LittleEndian, item.Bytes())
+	for _, item := range gr.items { // N
+		binary.Write(buf, binary.LittleEndian, item.Bytes()) // 33
 	}
 
+	// total: 47 + N*33
 	return buf.Bytes()
 }
